+++ storage/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "storage"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-trait = "0.1"
+serde = "1.0"
+slog = "2.5"
+tokio = { version = "1.7", features = ["sync"] }
