+++ storage/src/lib.rs
@@ -1,0 +1,5 @@
+mod hash_map_storage;
+mod storage;
+
+pub use hash_map_storage::HashMapStorage;
+pub use storage::{State, Storage};
